reorganize strategies
