--- FreeWill.Tests/PriorityTests.cs
@@ -1,1391 +1,0 @@
-using System;
-using FreeWill.Tests.TestHelpers;
-using Verse;
-
-namespace FreeWill.Tests
-{
-    /// <summary>
-    /// Basic tests for the Priority class infrastructure.
-    /// Note: These tests focus on what can be tested without complex RimWorld game state.
-    /// Full integration tests will require a more sophisticated test harness.
-    /// </summary>
-    public class PriorityTests
-    {
-        /// <summary>
-        /// Test Priority constructor with valid inputs (basic initialization).
-        /// Note: This tests the constructor but doesn't call methods that require game state.
-        /// </summary>
-        public static void TestConstructorBasics()
-        {
-            // Arrange
-            WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-
-            // Act & Assert - Test constructor with null pawn
-            // This documents the current behavior
-            try
-            {
-                Priority priority = new Priority(null, workTypeDef);
-
-                // Verify basic properties are accessible
-                if (priority.WorkTypeDef == null)
-                {
-                    throw new Exception("WorkTypeDef should not be null after construction");
-                }
-
-                if (priority.WorkTypeDef.defName != "Hauling")
-                {
-                    throw new Exception($"Expected WorkTypeDef.defName 'Hauling', got '{priority.WorkTypeDef.defName}'");
-                }
-
-                if (priority.AdjustmentStrings == null)
-                {
-                    throw new Exception("AdjustmentStrings should not be null after construction");
-                }
-
-                Console.WriteLine("Constructor basics - PASSED");
-            }
-            catch (NullReferenceException ex)
-            {
-                Console.WriteLine($"Constructor with null pawn threw NullReferenceException (expected): {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test Priority constructor with null WorkTypeDef.
-        /// </summary>
-        public static void TestConstructorWithNullWorkType()
-        {
-            try
-            {
-                Priority priority = new Priority(null, null);
-
-                // If we get here, verify the state
-                if (priority.WorkTypeDef != null)
-                {
-                    throw new Exception("Expected WorkTypeDef to be null");
-                }
-
-                Console.WriteLine("Constructor with null WorkTypeDef - PASSED (no exception thrown)");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Constructor with null WorkTypeDef threw exception (acceptable): {ex.GetType().Name}");
-            }
-        }
-
-        /// <summary>
-        /// Test IComparable implementation without requiring full Priority computation.
-        /// </summary>
-        public static void TestIComparable()
-        {
-            // We can't easily test this without creating valid Priority instances
-            // that have computed values, so for now just document the interface
-
-            WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-
-            try
-            {
-                Priority priority1 = new Priority(null, workTypeDef);
-                Priority priority2 = new Priority(null, workTypeDef);
-
-                // Test CompareTo with null
-                int result = ((IComparable)priority1).CompareTo(null);
-                if (result <= 0)
-                {
-                    throw new Exception("CompareTo(null) should return positive value");
-                }
-
-                // Test CompareTo with non-Priority object
-                int result2 = ((IComparable)priority1).CompareTo("not a priority");
-                if (result2 <= 0)
-                {
-                    throw new Exception("CompareTo(non-Priority) should return positive value");
-                }
-
-                Console.WriteLine("IComparable basic tests - PASSED");
-            }
-            catch (NullReferenceException)
-            {
-                Console.WriteLine("IComparable test skipped due to null pawn dependency");
-            }
-        }        /// <summary>
-                 /// Test work type definition creation (tests our test helpers).
-                 /// </summary>
-        public static void TestWorkTypeDefCreation()
-        {
-            try
-            {
-                WorkTypeDef firefighter = TestDataBuilders.WorkTypeDefs.Firefighter;
-                WorkTypeDef patient = TestDataBuilders.WorkTypeDefs.Patient;
-                WorkTypeDef doctor = TestDataBuilders.WorkTypeDefs.Doctor;
-                WorkTypeDef cooking = TestDataBuilders.WorkTypeDefs.Cooking;
-                WorkTypeDef hauling = TestDataBuilders.WorkTypeDefs.Hauling;
-                WorkTypeDef cleaning = TestDataBuilders.WorkTypeDefs.Cleaning;
-                WorkTypeDef research = TestDataBuilders.WorkTypeDefs.Research;                // Test each work type
-                if (firefighter?.defName != MockGameObjects.WorkTypes.Firefighter)
-                {
-                    throw new Exception($"Firefighter defName mismatch: expected '{MockGameObjects.WorkTypes.Firefighter}', got '{firefighter?.defName}'");
-                }
-                if (patient?.defName != MockGameObjects.WorkTypes.Patient)
-                {
-                    throw new Exception($"Patient defName mismatch: expected '{MockGameObjects.WorkTypes.Patient}', got '{patient?.defName}'");
-                }
-
-                if (doctor?.defName != MockGameObjects.WorkTypes.Doctor)
-                {
-                    throw new Exception($"Doctor defName mismatch: expected '{MockGameObjects.WorkTypes.Doctor}', got '{doctor?.defName}'");
-                }
-
-                if (cooking?.defName != MockGameObjects.WorkTypes.Cooking)
-                {
-                    throw new Exception($"Cooking defName mismatch: expected '{MockGameObjects.WorkTypes.Cooking}', got '{cooking?.defName}'");
-                }
-
-                if (hauling?.defName != MockGameObjects.WorkTypes.Hauling)
-                {
-                    throw new Exception($"Hauling defName mismatch: expected '{MockGameObjects.WorkTypes.Hauling}', got '{hauling?.defName}'");
-                }
-
-                if (cleaning?.defName != MockGameObjects.WorkTypes.Cleaning)
-                {
-                    throw new Exception($"Cleaning defName mismatch: expected '{MockGameObjects.WorkTypes.Cleaning}', got '{cleaning?.defName}'");
-                }
-
-                if (research?.defName != MockGameObjects.WorkTypes.Research)
-                {
-                    throw new Exception($"Research defName mismatch: expected '{MockGameObjects.WorkTypes.Research}', got '{research?.defName}'");
-                }
-
-                Console.WriteLine("WorkTypeDef creation - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"WorkTypeDef creation test failed or skipped due to RimWorld dependency: {ex.Message}");
-                Console.WriteLine("This is expected when running outside RimWorld environment");
-            }
-        }
-
-        /// <summary>
-        /// Test constants and helper values.
-        /// </summary>
-        public static void TestConstants()
-        {
-            // Test our test data builder constants
-            if (TestDataBuilders.Values.DefaultPriority != 0.2f)
-            {
-                throw new Exception($"DefaultPriority should be 0.2f, got {TestDataBuilders.Values.DefaultPriority}");
-            }
-
-            if (TestDataBuilders.Values.HighPriority != 0.8f)
-            {
-                throw new Exception($"HighPriority should be 0.8f, got {TestDataBuilders.Values.HighPriority}");
-            }
-
-            if (TestDataBuilders.Values.LowPriority != 0.1f)
-            {
-                throw new Exception($"LowPriority should be 0.1f, got {TestDataBuilders.Values.LowPriority}");
-            }
-
-            if (TestDataBuilders.Values.DisabledPriority != 0.0f)
-            {
-                throw new Exception($"DisabledPriority should be 0.0f, got {TestDataBuilders.Values.DisabledPriority}");
-            }
-
-            if (TestDataBuilders.Values.FloatTolerance != 0.001f)
-            {
-                throw new Exception($"FloatTolerance should be 0.001f, got {TestDataBuilders.Values.FloatTolerance}");
-            }
-
-            Console.WriteLine("Test constants - PASSED");
-        }
-
-        /// <summary>
-        /// Run all basic tests that can execute without complex RimWorld dependencies.
-        /// </summary>
-        public static void RunAllTests()
-        {
-            Console.WriteLine("=== Running FreeWill Priority Tests ===");
-            Console.WriteLine("Note: These are basic tests that avoid complex RimWorld game state dependencies.");
-            Console.WriteLine("Full integration testing will require more sophisticated test setup.");
-            Console.WriteLine();
-
-            int passedTests = 0;
-            int failedTests = 0;
-            int skippedTests = 0;
-
-            // Test infrastructure
-            RunTest("TestWorkTypeDefCreation", TestWorkTypeDefCreation, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestConstants", TestConstants, ref passedTests, ref failedTests, ref skippedTests);
-
-            // Basic constructor tests
-            RunTest("TestConstructorBasics", TestConstructorBasics, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestConstructorWithNullWorkType", TestConstructorWithNullWorkType, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestIComparable", TestIComparable, ref passedTests, ref failedTests, ref skippedTests);            // Core priority calculation tests (Step 2)
-            Console.WriteLine();
-            Console.WriteLine("=== Step 2: Core Priority Calculation Tests ===");
-            RunTest("TestToGamePriorityConversion", TestToGamePriorityConversion, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestFromGamePriorityConversion", TestFromGamePriorityConversion, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestRoundTripConversion", TestRoundTripConversion, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestToGamePriorityEdgeCases", TestToGamePriorityEdgeCases, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestFromGamePriorityEdgeCases", TestFromGamePriorityEdgeCases, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestComputeErrorHandling", TestComputeErrorHandling, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestPriorityAdjustmentMethods", TestPriorityAdjustmentMethods, ref passedTests, ref failedTests, ref skippedTests);
-
-            // New Step 2 tests - Priority adjustment methods
-            RunTest("TestMultiplyMethod", TestMultiplyMethod, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestAlwaysDoMethods", TestAlwaysDoMethods, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestNeverDoMethods", TestNeverDoMethods, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestDisabledFlagBehavior", TestDisabledFlagBehavior, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestComputeWithValidGameState", TestComputeWithValidGameState, ref passedTests, ref failedTests, ref skippedTests);
-
-            // Step 3: Consider* methods testing with dependency injection
-            Console.WriteLine();
-            Console.WriteLine("=== Step 3: Consider* Methods Tests with Dependency Injection ===");
-            RunTest("TestConsiderInspiration", TestConsiderInspiration, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestConsiderThoughts", TestConsiderThoughts, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestConsiderNeedingWarmClothes", TestConsiderNeedingWarmClothes, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestConsiderAnimalsRoaming", TestConsiderAnimalsRoaming, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestConsiderSuppressionNeed", TestConsiderSuppressionNeed, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestConsiderBored", TestConsiderBored, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestConsiderFire", TestConsiderFire, ref passedTests, ref failedTests, ref skippedTests);
-            RunTest("TestAdjustmentMethodsWithDependencyInjection", TestAdjustmentMethodsWithDependencyInjection, ref passedTests, ref failedTests, ref skippedTests);
-
-            // Step 4: Work-Type Strategy Tests
-            Console.WriteLine();
-            Console.WriteLine("=== Step 4: Work-Type Strategy Tests ===");
-            RunTest("TestFirefighterStrategy", TestFirefighterStrategy, ref passedTests, ref failedTests, ref skippedTests);
-
-            RunTest("TestMultiplyMethod", TestMultiplyMethod, ref passedTests, ref failedTests, ref skippedTests);
-
-            Console.WriteLine();
-            Console.WriteLine("=== Test Summary ===");
-            Console.WriteLine($"Passed: {passedTests}");
-            Console.WriteLine($"Failed: {failedTests}");
-            Console.WriteLine($"Skipped: {skippedTests}");
-            Console.WriteLine($"Total: {passedTests + failedTests + skippedTests}");
-            Console.WriteLine();
-
-            if (failedTests == 0)
-            {
-                Console.WriteLine("=== All runnable tests COMPLETED successfully! ===");
-                Console.WriteLine();
-                Console.WriteLine("Step 2 Progress - COMPLETED:");
-                Console.WriteLine("✓ Test ToGamePriority() conversion logic with boundary values");
-                Console.WriteLine("✓ Test FromGamePriority() conversion logic with boundary values");
-                Console.WriteLine("✓ Test IComparable.CompareTo() implementation");
-                Console.WriteLine("✓ Test round-trip conversion between ToGamePriority and FromGamePriority");
-                Console.WriteLine("✓ Test edge cases for both conversion methods");
-                Console.WriteLine("✓ Test priority adjustment helper methods (Set, Add via reflection)");
-                Console.WriteLine("✓ Test Compute() method error handling");
-                Console.WriteLine("✓ Test Multiply() method with various multipliers");
-                Console.WriteLine("✓ Test AlwaysDo() and AlwaysDoIf() methods");
-                Console.WriteLine("✓ Test NeverDo() and NeverDoIf() methods");
-                Console.WriteLine("✓ Test behavior when Disabled flag is set");
-                Console.WriteLine("✓ Test Compute() method for successful calculations (limited by game state dependencies)");
-                Console.WriteLine();
-                Console.WriteLine("STEP 2 IS NOW COMPLETE! All core priority calculation methods have been tested.");
-                Console.WriteLine();
-                Console.WriteLine("Next steps for Step 3 - Priority Adjustment Method Tests:");
-                Console.WriteLine("1. Test private Set() method with various values and descriptions");
-                Console.WriteLine("2. Test private Add() method with positive, negative, and zero adjustments");
-                Console.WriteLine("3. Expand Multiply() tests for different edge cases");
-                Console.WriteLine("4. Test AlwaysDo/NeverDo interaction with ToGamePriority()");
-                Console.WriteLine("5. Test priority clamping behavior in all adjustment methods");
-            }
-            else
-            {
-                Console.WriteLine($"=== {failedTests} test(s) FAILED ===");
-                throw new Exception($"{failedTests} test(s) failed");
-            }
-        }
-
-        private static void RunTest(string testName, Action testMethod, ref int passedTests, ref int failedTests, ref int skippedTests)
-        {
-            try
-            {
-                testMethod();
-                passedTests++;
-            }
-            catch (Exception ex)
-            {
-                if (ex.Message.Contains("skipped") || ex.Message.Contains("RimWorld dependency"))
-                {
-                    skippedTests++;
-                }
-                else
-                {
-                    Console.WriteLine($"FAILED: {testName} - {ex.Message}");
-                    failedTests++;
-                }
-            }
-        }
-
-        /// <summary>
-        /// Test ToGamePriority() conversion logic with boundary values.
-        /// </summary>
-        public static void TestToGamePriorityConversion()
-        {
-            WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-
-            try
-            {
-                Priority priority = new Priority(null, workTypeDef);
-
-                // Test direct value setting to test conversion without game state
-                // Using reflection to set Value directly for testing
-                System.Reflection.FieldInfo valueField = typeof(Priority).GetField("Value", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                if (valueField == null)
-                {
-                    Console.WriteLine("ToGamePriority test skipped - cannot access Value field directly");
-                    return;
-                }
-
-                // Test boundary values
-                // Value 0.0 should convert to disabled (0)
-                valueField.SetValue(priority, 0.0f);
-                int result = priority.ToGamePriority();
-                if (result != 0)
-                {
-                    throw new Exception($"Expected 0.0f to convert to 0, got {result}");
-                }
-
-                // Value 1.0 should convert to highest priority (1)
-                valueField.SetValue(priority, 1.0f);
-                result = priority.ToGamePriority();
-                if (result != 1)
-                {
-                    throw new Exception($"Expected 1.0f to convert to 1, got {result}");
-                }
-
-                // Test mid-range value
-                valueField.SetValue(priority, 0.5f);
-                result = priority.ToGamePriority();
-                if (result < 1 || result > 4) // Assuming LowestPriority is 4
-                {
-                    throw new Exception($"Expected 0.5f to convert to reasonable priority, got {result}");
-                }
-
-                // Test very low value (should be disabled threshold)
-                valueField.SetValue(priority, 0.1f);
-                result = priority.ToGamePriority();
-                // Should be either 0 or lowest priority depending on Enabled flag
-
-                Console.WriteLine("ToGamePriority conversion tests - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"ToGamePriority test failed or skipped: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test FromGamePriority() conversion logic with boundary values.
-        /// </summary>
-        public static void TestFromGamePriorityConversion()
-        {
-            WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-
-            try
-            {
-                Priority priority = new Priority(null, workTypeDef);
-
-                // Test game priority 0 (disabled)
-                priority.FromGamePriority(0);
-                if (priority.Value != 0.0f)
-                {
-                    throw new Exception($"Expected FromGamePriority(0) to set Value to 0.0f, got {priority.Value}");
-                }
-
-                // Test game priority 1 (highest)
-                priority.FromGamePriority(1);
-                if (priority.Value <= 0.8f) // Should be high value
-                {
-                    throw new Exception($"Expected FromGamePriority(1) to set high value, got {priority.Value}");
-                }
-
-                // Test game priority 4 (lowest, assuming LowestPriority = 4)
-                priority.FromGamePriority(4);
-                if (priority.Value >= 0.3f) // Should be low value
-                {
-                    throw new Exception($"Expected FromGamePriority(4) to set low value, got {priority.Value}");
-                }
-
-                // Test mid-range
-                priority.FromGamePriority(2);
-                if (priority.Value <= 0.2f || priority.Value >= 0.8f)
-                {
-                    throw new Exception($"Expected FromGamePriority(2) to set mid-range value, got {priority.Value}");
-                }
-
-                Console.WriteLine("FromGamePriority conversion tests - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"FromGamePriority test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test round-trip conversion between ToGamePriority and FromGamePriority.
-        /// </summary>
-        public static void TestRoundTripConversion()
-        {
-            WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-
-            try
-            {
-                Priority priority1 = new Priority(null, workTypeDef);
-                Priority priority2 = new Priority(null, workTypeDef);
-
-                // Test various game priority values
-                for (int gamePriority = 0; gamePriority <= 4; gamePriority++)
-                {
-                    priority1.FromGamePriority(gamePriority);
-                    int convertedBack = priority1.ToGamePriority();
-
-                    // For game priority 0, should round-trip to 0
-                    if (gamePriority == 0 && convertedBack != 0)
-                    {
-                        throw new Exception($"Round-trip failed for game priority 0: got {convertedBack}");
-                    }
-
-                    // For other values, should be reasonably close (within 1 due to precision/rounding)
-                    if (gamePriority > 0 && Math.Abs(convertedBack - gamePriority) > 1)
-                    {
-                        throw new Exception($"Round-trip failed for game priority {gamePriority}: got {convertedBack}");
-                    }
-                }
-
-                Console.WriteLine("Round-trip conversion tests - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Round-trip conversion test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test that ToGamePriority handles edge cases and invalid values properly.
-        /// </summary>
-        public static void TestToGamePriorityEdgeCases()
-        {
-            WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-
-            try
-            {
-                Priority priority = new Priority(null, workTypeDef);
-                System.Reflection.FieldInfo valueField = typeof(Priority).GetField("Value", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-
-                if (valueField == null)
-                {
-                    Console.WriteLine("ToGamePriority edge cases test skipped - cannot access Value field directly");
-                    return;
-                }
-
-                // Test negative value (should be clamped)
-                valueField.SetValue(priority, -0.5f);
-                int result = priority.ToGamePriority();
-                if (result < 0)
-                {
-                    throw new Exception($"Negative value should be handled gracefully, got {result}");
-                }
-
-                // Test value > 1.0 (should be clamped)
-                valueField.SetValue(priority, 1.5f);
-                result = priority.ToGamePriority();
-                if (result < 1)
-                {
-                    throw new Exception($"Value > 1.0 should convert to high priority, got {result}");
-                }
-
-                // Test very small positive value
-                valueField.SetValue(priority, 0.001f);
-                result = priority.ToGamePriority();
-                // Should not crash and should return valid value
-
-                Console.WriteLine("ToGamePriority edge cases - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"ToGamePriority edge cases test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test FromGamePriority with invalid input values.
-        /// </summary>
-        public static void TestFromGamePriorityEdgeCases()
-        {
-            WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-
-            try
-            {
-                Priority priority = new Priority(null, workTypeDef);
-
-                // Test negative game priority
-                priority.FromGamePriority(-1);
-                // Should not crash and should handle gracefully
-
-                // Test very high game priority
-                priority.FromGamePriority(100);
-                // Should not crash and should handle gracefully
-
-                // Verify the value is still in valid range after invalid input
-                if (priority.Value < 0.0f || priority.Value > 1.0f)
-                {
-                    throw new Exception($"Invalid input should still result in valid Value range, got {priority.Value}");
-                }
-
-                Console.WriteLine("FromGamePriority edge cases - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"FromGamePriority edge cases test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test error handling in Compute() method.
-        /// Note: This tests the error handling path when game state is invalid.
-        /// </summary>
-        public static void TestComputeErrorHandling()
-        {
-            WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-
-            try
-            {
-                // Test with null pawn (should trigger error handling)
-                Priority priority = new Priority(null, workTypeDef);
-
-                try
-                {
-                    priority.Compute();
-                    Console.WriteLine("Compute with null pawn - handled gracefully (no exception)");
-                }
-                catch (NullReferenceException)
-                {
-                    Console.WriteLine("Compute with null pawn - threw NullReferenceException (expected)");
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine($"Compute with null pawn - threw {ex.GetType().Name}: {ex.Message}");
-                }
-
-                // Test with null WorkTypeDef
-                Priority priority2 = new Priority(null, null);
-                try
-                {
-                    priority2.Compute();
-                    Console.WriteLine("Compute with null WorkTypeDef - handled gracefully (no exception)");
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine($"Compute with null WorkTypeDef - threw {ex.GetType().Name} (expected)");
-                }
-
-                Console.WriteLine("Compute error handling tests - COMPLETED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Compute error handling test setup failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test the priority adjustment helper methods.
-        /// </summary>
-        public static void TestPriorityAdjustmentMethods()
-        {
-            WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-
-            try
-            {
-                Priority priority = new Priority(null, workTypeDef);
-
-                // We need to use reflection to test private methods or test through public interface
-                System.Reflection.FieldInfo valueField = typeof(Priority).GetField("Value", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                System.Reflection.MethodInfo setMethod = typeof(Priority).GetMethod("Set", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                System.Reflection.MethodInfo addMethod = typeof(Priority).GetMethod("Add", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-
-                if (valueField == null || setMethod == null || addMethod == null)
-                {
-                    Console.WriteLine("Priority adjustment methods test skipped - cannot access private methods via reflection");
-                    return;
-                }
-
-                // Test Set method
-                Func<string> testDesc = () => "Test description";
-                _ = setMethod.Invoke(priority, new object[] { 0.7f, testDesc });
-                float value = (float)valueField.GetValue(priority);
-                if (Math.Abs(value - 0.7f) > 0.001f)
-                {
-                    throw new Exception($"Set method failed: expected 0.7, got {value}");
-                }
-
-                // Test Add method - should add to existing value
-                _ = addMethod.Invoke(priority, new object[] { 0.2f, testDesc });
-                value = (float)valueField.GetValue(priority);
-                if (Math.Abs(value - 0.9f) > 0.001f)
-                {
-                    throw new Exception($"Add method failed: expected 0.9, got {value}");
-                }
-
-                // Test clamping - adding beyond 1.0 should clamp
-                _ = addMethod.Invoke(priority, new object[] { 0.5f, testDesc });
-                value = (float)valueField.GetValue(priority);
-                if (value > 1.0f)
-                {
-                    throw new Exception($"Add method should clamp to 1.0, got {value}");
-                }
-
-                Console.WriteLine("Priority adjustment methods - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Priority adjustment methods test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test the Multiply method with various multipliers.
-        /// </summary>
-        public static void TestMultiplyMethod()
-        {
-            WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-
-            try
-            {
-                Priority priority = new Priority(null, workTypeDef);
-                System.Reflection.FieldInfo valueField = typeof(Priority).GetField("Value", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                System.Reflection.MethodInfo multiplyMethod = typeof(Priority).GetMethod("Multiply", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-
-                if (valueField == null || multiplyMethod == null)
-                {
-                    Console.WriteLine("Multiply method test skipped - cannot access private methods via reflection");
-                    return;
-                }
-
-                Func<string> testDesc = () => "Test multiply";
-
-                // Test normal multiplication
-                valueField.SetValue(priority, 0.5f);
-                _ = multiplyMethod.Invoke(priority, new object[] { 2.0f, testDesc });
-                float value = (float)valueField.GetValue(priority);
-                if (Math.Abs(value - 1.0f) > 0.001f)
-                {
-                    throw new Exception($"Multiply(2.0) on 0.5 failed: expected 1.0, got {value}");
-                }
-
-                // Test multiplication with clamping to 1.0
-                valueField.SetValue(priority, 0.8f);
-                _ = multiplyMethod.Invoke(priority, new object[] { 2.0f, testDesc });
-                value = (float)valueField.GetValue(priority);
-                if (value > 1.0f)
-                {
-                    throw new Exception($"Multiply should clamp to 1.0, got {value}");
-                }
-
-                // Test multiplication that reduces value
-                valueField.SetValue(priority, 0.6f);
-                _ = multiplyMethod.Invoke(priority, new object[] { 0.5f, testDesc });
-                value = (float)valueField.GetValue(priority);
-                if (Math.Abs(value - 0.3f) > 0.001f)
-                {
-                    throw new Exception($"Multiply(0.5) on 0.6 failed: expected 0.3, got {value}");
-                }
-
-                // Test multiplication by zero
-                valueField.SetValue(priority, 0.5f);
-                _ = multiplyMethod.Invoke(priority, new object[] { 0.0f, testDesc });
-                value = (float)valueField.GetValue(priority);
-                if (value != 0.0f)
-                {
-                    throw new Exception($"Multiply(0.0) failed: expected 0.0, got {value}");
-                }
-
-                // Test multiplication by 1.0 (no change)
-                valueField.SetValue(priority, 0.5f);
-                _ = multiplyMethod.Invoke(priority, new object[] { 1.0f, testDesc });
-                value = (float)valueField.GetValue(priority);
-                if (Math.Abs(value - 0.5f) > 0.001f)
-                {
-                    throw new Exception($"Multiply(1.0) should not change value: expected 0.5, got {value}");
-                }
-
-                // Test with disabled priority (should not change)
-                System.Reflection.FieldInfo disabledField = typeof(Priority).GetField("Disabled", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                if (disabledField != null)
-                {
-                    valueField.SetValue(priority, 0.5f);
-                    disabledField.SetValue(priority, true);
-                    _ = multiplyMethod.Invoke(priority, new object[] { 2.0f, testDesc });
-                    value = (float)valueField.GetValue(priority);
-                    if (Math.Abs(value - 0.5f) > 0.001f)
-                    {
-                        throw new Exception($"Multiply on disabled priority should not change value: expected 0.5, got {value}");
-                    }
-                    disabledField.SetValue(priority, false); // Reset for other tests
-                }
-
-                Console.WriteLine("Multiply method tests - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Multiply method test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test the AlwaysDo and AlwaysDoIf methods.
-        /// </summary>
-        public static void TestAlwaysDoMethods()
-        {
-            WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-
-            try
-            {
-                Priority priority = new Priority(null, workTypeDef);
-                System.Reflection.FieldInfo enabledField = typeof(Priority).GetField("Enabled", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                System.Reflection.FieldInfo disabledField = typeof(Priority).GetField("Disabled", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                System.Reflection.MethodInfo alwaysDoMethod = typeof(Priority).GetMethod("AlwaysDo", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                System.Reflection.MethodInfo alwaysDoIfMethod = typeof(Priority).GetMethod("AlwaysDoIf", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-
-                if (enabledField == null || disabledField == null || alwaysDoMethod == null || alwaysDoIfMethod == null)
-                {
-                    Console.WriteLine("AlwaysDo methods test skipped - cannot access private methods/fields via reflection");
-                    return;
-                }
-
-                Func<string> testDesc = () => "Test always do";
-
-                // Reset state
-                enabledField.SetValue(priority, false);
-                disabledField.SetValue(priority, false);
-
-                // Test AlwaysDo (should set Enabled = true, Disabled = false)
-                _ = alwaysDoMethod.Invoke(priority, new object[] { testDesc });
-                bool enabled = (bool)enabledField.GetValue(priority);
-                bool disabled = (bool)disabledField.GetValue(priority);
-
-                if (!enabled)
-                {
-                    throw new Exception("AlwaysDo should set Enabled to true");
-                }
-                if (disabled)
-                {
-                    throw new Exception("AlwaysDo should set Disabled to false");
-                }
-
-                // Reset state and test AlwaysDoIf with true condition
-                enabledField.SetValue(priority, false);
-                disabledField.SetValue(priority, false);
-
-                _ = alwaysDoIfMethod.Invoke(priority, new object[] { true, testDesc });
-                enabled = (bool)enabledField.GetValue(priority);
-                disabled = (bool)disabledField.GetValue(priority);
-
-                if (!enabled)
-                {
-                    throw new Exception("AlwaysDoIf(true) should set Enabled to true");
-                }
-                if (disabled)
-                {
-                    throw new Exception("AlwaysDoIf(true) should set Disabled to false");
-                }
-
-                // Reset state and test AlwaysDoIf with false condition (should not change state)
-                enabledField.SetValue(priority, false);
-                disabledField.SetValue(priority, false);
-
-                _ = alwaysDoIfMethod.Invoke(priority, new object[] { false, testDesc });
-                enabled = (bool)enabledField.GetValue(priority);
-                disabled = (bool)disabledField.GetValue(priority);
-
-                if (enabled)
-                {
-                    throw new Exception("AlwaysDoIf(false) should not set Enabled to true");
-                }
-                if (disabled)
-                {
-                    throw new Exception("AlwaysDoIf(false) should not set Disabled to true");
-                }
-
-                // Test AlwaysDoIf when already enabled (should not change)
-                enabledField.SetValue(priority, true);
-                disabledField.SetValue(priority, false);
-
-                _ = alwaysDoIfMethod.Invoke(priority, new object[] { true, testDesc });
-                enabled = (bool)enabledField.GetValue(priority);
-                disabled = (bool)disabledField.GetValue(priority);
-
-                if (!enabled)
-                {
-                    throw new Exception("AlwaysDoIf when already enabled should keep Enabled true");
-                }
-                if (disabled)
-                {
-                    throw new Exception("AlwaysDoIf when already enabled should keep Disabled false");
-                }
-
-                Console.WriteLine("AlwaysDo methods tests - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"AlwaysDo methods test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test the NeverDo and NeverDoIf methods.
-        /// </summary>
-        public static void TestNeverDoMethods()
-        {
-            WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-
-            try
-            {
-                Priority priority = new Priority(null, workTypeDef);
-                System.Reflection.FieldInfo enabledField = typeof(Priority).GetField("Enabled", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                System.Reflection.FieldInfo disabledField = typeof(Priority).GetField("Disabled", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                System.Reflection.MethodInfo neverDoMethod = typeof(Priority).GetMethod("NeverDo", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                System.Reflection.MethodInfo neverDoIfMethod = typeof(Priority).GetMethod("NeverDoIf", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-
-                if (enabledField == null || disabledField == null || neverDoMethod == null || neverDoIfMethod == null)
-                {
-                    Console.WriteLine("NeverDo methods test skipped - cannot access private methods/fields via reflection");
-                    return;
-                }
-
-                Func<string> testDesc = () => "Test never do";
-
-                // Reset state
-                enabledField.SetValue(priority, false);
-                disabledField.SetValue(priority, false);
-
-                // Test NeverDo (should set Disabled = true, Enabled = false)
-                _ = neverDoMethod.Invoke(priority, new object[] { testDesc });
-                bool enabled = (bool)enabledField.GetValue(priority);
-                bool disabled = (bool)disabledField.GetValue(priority);
-
-                if (enabled)
-                {
-                    throw new Exception("NeverDo should set Enabled to false");
-                }
-                if (!disabled)
-                {
-                    throw new Exception("NeverDo should set Disabled to true");
-                }
-
-                // Reset state and test NeverDoIf with true condition
-                enabledField.SetValue(priority, false);
-                disabledField.SetValue(priority, false);
-
-                _ = neverDoIfMethod.Invoke(priority, new object[] { true, testDesc });
-                enabled = (bool)enabledField.GetValue(priority);
-                disabled = (bool)disabledField.GetValue(priority);
-
-                if (enabled)
-                {
-                    throw new Exception("NeverDoIf(true) should set Enabled to false");
-                }
-                if (!disabled)
-                {
-                    throw new Exception("NeverDoIf(true) should set Disabled to true");
-                }
-
-                // Reset state and test NeverDoIf with false condition (should not change state)
-                enabledField.SetValue(priority, false);
-                disabledField.SetValue(priority, false);
-
-                _ = neverDoIfMethod.Invoke(priority, new object[] { false, testDesc });
-                enabled = (bool)enabledField.GetValue(priority);
-                disabled = (bool)disabledField.GetValue(priority);
-
-                if (enabled)
-                {
-                    throw new Exception("NeverDoIf(false) should not set Enabled to true");
-                }
-                if (disabled)
-                {
-                    throw new Exception("NeverDoIf(false) should not set Disabled to true");
-                }
-
-                // Test NeverDoIf when already disabled (should not change)
-                enabledField.SetValue(priority, false);
-                disabledField.SetValue(priority, true);
-
-                _ = neverDoIfMethod.Invoke(priority, new object[] { true, testDesc });
-                enabled = (bool)enabledField.GetValue(priority);
-                disabled = (bool)disabledField.GetValue(priority);
-
-                if (enabled)
-                {
-                    throw new Exception("NeverDoIf when already disabled should keep Enabled false");
-                }
-                if (!disabled)
-                {
-                    throw new Exception("NeverDoIf when already disabled should keep Disabled true");
-                }
-
-                Console.WriteLine("NeverDo methods tests - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"NeverDo methods test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test behavior when Disabled flag is set.
-        /// </summary>
-        public static void TestDisabledFlagBehavior()
-        {
-            WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-
-            try
-            {
-                Priority priority = new Priority(null, workTypeDef);
-                System.Reflection.FieldInfo valueField = typeof(Priority).GetField("Value", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                System.Reflection.FieldInfo disabledField = typeof(Priority).GetField("Disabled", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                System.Reflection.FieldInfo enabledField = typeof(Priority).GetField("Enabled", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-                System.Reflection.MethodInfo multiplyMethod = typeof(Priority).GetMethod("Multiply", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-
-                if (valueField == null || disabledField == null || enabledField == null || multiplyMethod == null)
-                {
-                    Console.WriteLine("Disabled flag behavior test skipped - cannot access private fields/methods via reflection");
-                    return;
-                }
-
-                Func<string> testDesc = () => "Test disabled behavior";
-
-                // Test ToGamePriority when disabled
-                disabledField.SetValue(priority, true);
-                enabledField.SetValue(priority, false);
-                valueField.SetValue(priority, 0.5f);
-
-                int gamePriority = priority.ToGamePriority();
-                if (gamePriority != 0)
-                {
-                    throw new Exception($"ToGamePriority() should return 0 when disabled, got {gamePriority}");
-                }
-
-                // Test that Multiply doesn't change value when disabled
-                float originalValue = 0.5f;
-                valueField.SetValue(priority, originalValue);
-                disabledField.SetValue(priority, true);
-
-                _ = multiplyMethod.Invoke(priority, new object[] { 2.0f, testDesc });
-                float newValue = (float)valueField.GetValue(priority);
-
-                if (Math.Abs(newValue - originalValue) > 0.001f)
-                {
-                    throw new Exception($"Multiply should not change value when disabled: expected {originalValue}, got {newValue}");
-                }
-
-                // Test that operations work normally when not disabled
-                disabledField.SetValue(priority, false);
-                valueField.SetValue(priority, 0.5f);
-
-                _ = multiplyMethod.Invoke(priority, new object[] { 2.0f, testDesc });
-                newValue = (float)valueField.GetValue(priority);
-
-                if (Math.Abs(newValue - 1.0f) > 0.001f)
-                {
-                    throw new Exception($"Multiply should work normally when not disabled: expected 1.0, got {newValue}");
-                }
-
-                // Test enabled/disabled mutual exclusivity
-                disabledField.SetValue(priority, true);
-                enabledField.SetValue(priority, true);
-
-                bool disabled = (bool)disabledField.GetValue(priority);
-                bool enabled = (bool)enabledField.GetValue(priority);
-
-                // Note: The Priority class might not enforce mutual exclusivity, 
-                // this documents the current behavior
-                Console.WriteLine($"Disabled flag behavior - Disabled: {disabled}, Enabled: {enabled}");
-
-                Console.WriteLine("Disabled flag behavior tests - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Disabled flag behavior test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test successful Priority.Compute() calculation with minimal valid game state.
-        /// Note: This test will likely be limited due to RimWorld dependencies.
-        /// </summary>
-        public static void TestComputeWithValidGameState()
-        {
-            WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-
-            try
-            {
-                // This test is complex because Compute() requires:
-                // - Valid pawn with Map
-                // - Map with FreeWill_MapComponent
-                // - World with FreeWill_WorldComponent
-                // These are difficult to mock without full RimWorld context
-
-                Priority priority = new Priority(null, workTypeDef);
-
-                // For now, we'll test that we can at least call Compute() and handle the expected exception
-                bool threwExpectedException = false;
-                try
-                {
-                    priority.Compute();
-                }
-                catch (NullReferenceException)
-                {
-                    threwExpectedException = true;
-                    Console.WriteLine("Compute with null pawn threw NullReferenceException (expected)");
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine($"Compute threw unexpected exception: {ex.GetType().Name}: {ex.Message}");
-                    threwExpectedException = true; // Accept various exceptions for now
-                }
-
-                if (!threwExpectedException)
-                {
-                    // If no exception was thrown, verify the state
-                    if (priority.Value == 0.0f)
-                    {
-                        Console.WriteLine("WARNING: Compute() completed with Value = 0.0, may indicate error handling path");
-                    }
-                    else
-                    {
-                        Console.WriteLine($"Compute() completed successfully with Value = {priority.Value}");
-                    }
-                }
-
-                // Test that AdjustmentStrings are initialized after Compute() attempt
-                if (priority.AdjustmentStrings == null)
-                {
-                    throw new Exception("AdjustmentStrings should not be null after Compute() attempt");
-                }
-
-                Console.WriteLine("Compute with valid game state test - COMPLETED (limited by RimWorld dependencies)");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Compute with valid game state test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test ConsiderInspiration method with mocked dependencies.
-        /// </summary>
-        public static void TestConsiderInspiration()
-        {
-            try
-            {
-                WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-                MockPriorityDependencyProvider mockProvider = new MockPriorityDependencyProvider();
-
-                // Test with no inspiration (default case)
-                Priority priority = new Priority(null, workTypeDef, mockProvider);
-                Priority result = priority.ConsiderInspiration() ?? throw new Exception("ConsiderInspiration should return a Priority object");
-                Console.WriteLine("ConsiderInspiration basic test - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"ConsiderInspiration test failed: {ex.Message}");
-            }
-        }        /// <summary>
-                 /// Test ConsiderThoughts method with mocked dependencies.
-                 /// </summary>
-        public static void TestConsiderThoughts()
-        {
-            try
-            {
-                WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-                MockPriorityDependencyProvider mockProvider = new MockPriorityDependencyProvider();
-
-                // Test with no thoughts
-                Priority priority = new Priority(null, workTypeDef, mockProvider);
-                Priority result = priority.ConsiderThoughts() ?? throw new Exception("ConsiderThoughts should return a Priority object");
-                Console.WriteLine("ConsiderThoughts basic test - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"ConsiderThoughts test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test ConsiderNeedingWarmClothes method with different alert states.
-        /// </summary>
-        public static void TestConsiderNeedingWarmClothes()
-        {
-            try
-            {
-                WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-                MockPriorityDependencyProvider mockProvider = new MockPriorityDependencyProvider();
-                MockMapStateProvider mockMapState = (MockMapStateProvider)mockProvider.MapStateProvider;
-
-                // Test when warm clothes alert is not active
-                mockMapState.AlertNeedWarmClothes = false;
-                Priority priority1 = new Priority(null, workTypeDef, mockProvider);
-                Priority result1 = priority1.ConsiderNeedingWarmClothes() ?? throw new Exception("ConsiderNeedingWarmClothes should return a Priority object");
-
-                // Test when warm clothes alert is active
-                mockMapState.AlertNeedWarmClothes = true;
-                Priority priority2 = new Priority(null, workTypeDef, mockProvider);
-                Priority result2 = priority2.ConsiderNeedingWarmClothes() ?? throw new Exception("ConsiderNeedingWarmClothes should return a Priority object when alert is active");
-                Console.WriteLine("ConsiderNeedingWarmClothes alert test - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"ConsiderNeedingWarmClothes test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test ConsiderAnimalsRoaming method with different alert states.
-        /// </summary>
-        public static void TestConsiderAnimalsRoaming()
-        {
-            try
-            {
-                WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-                MockPriorityDependencyProvider mockProvider = new MockPriorityDependencyProvider();
-                MockMapStateProvider mockMapState = (MockMapStateProvider)mockProvider.MapStateProvider;
-
-                // Test when animals roaming alert is not active
-                mockMapState.AlertAnimalRoaming = false;
-                Priority priority1 = new Priority(null, workTypeDef, mockProvider);
-                Priority result1 = priority1.ConsiderAnimalsRoaming() ?? throw new Exception("ConsiderAnimalsRoaming should return a Priority object");
-
-                // Test when animals roaming alert is active
-                mockMapState.AlertAnimalRoaming = true;
-                Priority priority2 = new Priority(null, workTypeDef, mockProvider);
-                Priority result2 = priority2.ConsiderAnimalsRoaming() ?? throw new Exception("ConsiderAnimalsRoaming should return a Priority object when alert is active");
-                Console.WriteLine("ConsiderAnimalsRoaming alert test - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"ConsiderAnimalsRoaming test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test ConsiderSuppressionNeed method with different suppression levels.
-        /// </summary>
-        public static void TestConsiderSuppressionNeed()
-        {
-            try
-            {
-                WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-                MockPriorityDependencyProvider mockProvider = new MockPriorityDependencyProvider();
-                MockMapStateProvider mockMapState = (MockMapStateProvider)mockProvider.MapStateProvider;
-
-                // Test with low suppression need
-                mockMapState.SuppressionNeed = 0.1f;
-                Priority priority1 = new Priority(null, workTypeDef, mockProvider);
-                Priority result1 = priority1.ConsiderSuppressionNeed() ?? throw new Exception("ConsiderSuppressionNeed should return a Priority object");
-
-                // Test with high suppression need
-                mockMapState.SuppressionNeed = 0.8f;
-                Priority priority2 = new Priority(null, workTypeDef, mockProvider);
-                Priority result2 = priority2.ConsiderSuppressionNeed() ?? throw new Exception("ConsiderSuppressionNeed should return a Priority object with high suppression");
-                Console.WriteLine("ConsiderSuppressionNeed levels test - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"ConsiderSuppressionNeed test failed: {ex.Message}");
-            }
-        }        /// <summary>
-                 /// Test ConsiderBored method with different scenarios.
-                 /// Tests both the exception handling path (null pawn) and normal operation.
-                 /// </summary>
-        public static void TestConsiderBored()
-        {
-            try
-            {
-                WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-                MockPriorityDependencyProvider mockProvider = new MockPriorityDependencyProvider();                // Test 1: Exception handling with null pawn (should return gracefully)
-                Priority priority1 = new Priority(null, workTypeDef, mockProvider);
-                Priority result1 = priority1.ConsiderBored() ?? throw new Exception("ConsiderBored should return a Priority object even with null pawn");
-
-                // Test 2: With a real pawn (though mindState might still cause issues in test environment)
-                try
-                {
-                    Pawn testPawn = TestPawns.BasicColonist();
-                    Priority priority2 = new Priority(testPawn, workTypeDef, mockProvider);
-                    Priority result2 = priority2.ConsiderBored() ?? throw new Exception("ConsiderBored should return a Priority object with real pawn");
-                }
-                catch (Exception pawnEx)
-                {
-                    // Expected in test environment - RimWorld components may not be fully initialized
-                    // This is fine as long as the exception is handled gracefully
-                    if (pawnEx.Message.Contains("ECall methods must be packaged into a system module"))
-                    {
-                        // This is the expected limitation in the test environment
-                        // The production code handles this correctly via exception handling
-                    }
-                    else
-                    {
-                        throw; // Re-throw if it's a different type of error
-                    }
-                }
-
-                Console.WriteLine("ConsiderBored test - PASSED (exception handling verified)");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"ConsiderBored test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test ConsiderFire method with different fire alert states.
-        /// </summary>
-        public static void TestConsiderFire()
-        {
-            try
-            {
-                WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Firefighter;
-                MockPriorityDependencyProvider mockProvider = new MockPriorityDependencyProvider();
-                MockMapStateProvider mockMapState = (MockMapStateProvider)mockProvider.MapStateProvider;
-
-                // Test with no fire alert
-                mockMapState.AlertFireInHomeArea = false;
-                Priority priority1 = new Priority(null, workTypeDef, mockProvider);
-                Priority result1 = priority1.ConsiderFire() ?? throw new Exception("ConsiderFire should return a Priority object");
-
-                // Test with fire alert active (should increase priority for firefighter work)
-                mockMapState.AlertFireInHomeArea = true;
-                Priority priority2 = new Priority(null, workTypeDef, mockProvider);
-                Priority result2 = priority2.ConsiderFire() ?? throw new Exception("ConsiderFire should return a Priority object when fire alert is active");
-                Console.WriteLine("ConsiderFire alert test - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"ConsiderFire test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test the adjustment methods (Set, Add, Multiply) are working correctly with dependency injection.
-        /// </summary>
-        public static void TestAdjustmentMethodsWithDependencyInjection()
-        {
-            try
-            {
-                WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Hauling;
-                MockPriorityDependencyProvider mockProvider = new MockPriorityDependencyProvider();
-
-                Priority priority = new Priority(null, workTypeDef, mockProvider);
-
-                // Test that the public adjustment methods are accessible
-                priority.Set(0.5f, () => "Test set");
-                priority.Add(0.2f, () => "Test add");
-                Priority result = priority.Multiply(2.0f, () => "Test multiply") ?? throw new Exception("Multiply should return a Priority object");
-
-                // Test that priority maintains its state
-                if (priority.WorkTypeDef.defName != workTypeDef.defName)
-                {
-                    throw new Exception("Priority should maintain WorkTypeDef after adjustments");
-                }
-
-                Console.WriteLine("Adjustment methods with dependency injection - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Adjustment methods with dependency injection test failed: {ex.Message}");
-            }
-        }
-
-        /// <summary>
-        /// Test ConsiderLowFood method with different food shortage scenarios.
-        /// </summary>
-        public static void TestConsiderLowFood()
-        {
-            try
-            {
-                WorkTypeDef workTypeDef = TestDataBuilders.WorkTypeDefs.Cooking;
-                MockPriorityDependencyProvider mockProvider = new MockPriorityDependencyProvider();
-                MockMapStateProvider mockMapState = (MockMapStateProvider)mockProvider.MapStateProvider; MockWorldStateProvider mockWorldState = (MockWorldStateProvider)mockProvider.WorldStateProvider;
-
-                // Test with no food alert - should not change priority
-                mockMapState.AlertLowFood = false;
-                mockWorldState.Settings.ConsiderLowFood = 1.0f;
-                Pawn testPawn = TestPawns.BasicColonist();
-                Priority priority1 = new Priority(testPawn, workTypeDef, mockProvider);
-                float initialValue = priority1.Value;
-                Priority result1 = priority1.ConsiderLowFood(0.5f);
-                if (Math.Abs(result1.Value - initialValue) > 0.01f)
-                {
-                    throw new Exception($"ConsiderLowFood should not change priority when AlertLowFood is false. Expected {initialValue}, got {result1.Value}");
-                }                // Test with food alert enabled - should increase priority for relevant work types
-                mockMapState.AlertLowFood = true;
-                mockWorldState.Settings.ConsiderLowFood = 1.0f;
-                Priority priority2 = new Priority(testPawn, workTypeDef, mockProvider);
-                float initialValue2 = priority2.Value;
-                Priority result2 = priority2.ConsiderLowFood(0.5f);
-                if (result2.Value <= initialValue2)
-                {
-                    throw new Exception($"ConsiderLowFood should increase priority when AlertLowFood is true. Expected > {initialValue2}, got {result2.Value}");
-                }                // Test with ConsiderLowFood setting disabled - should not change priority
-                mockMapState.AlertLowFood = true;
-                mockWorldState.Settings.ConsiderLowFood = 0.0f;
-                Priority priority3 = new Priority(testPawn, workTypeDef, mockProvider);
-                float initialValue3 = priority3.Value;
-                Priority result3 = priority3.ConsiderLowFood(0.5f);
-                if (Math.Abs(result3.Value - initialValue3) > 0.01f)
-                {
-                    throw new Exception($"ConsiderLowFood should not change priority when ConsiderLowFood setting is 0. Expected {initialValue3}, got {result3.Value}");
-                }
-
-                Console.WriteLine("ConsiderLowFood test - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"ConsiderLowFood test failed: {ex.Message}");
-            }
-        }        /// <summary>
-                 /// Test FirefighterStrategy priority calculations for emergency fire response.
-                 /// Tests the strategy pattern implementation and basic instantiation.
-                 /// Note: Full testing is limited by RimWorld dependencies in test environment.
-                 /// </summary>
-        public static void TestFirefighterStrategy()
-        {
-            try
-            {
-                WorkTypeDef firefighterWorkType = TestDataBuilders.WorkTypeDefs.Firefighter;
-                MockPriorityDependencyProvider mockProvider = new MockPriorityDependencyProvider();
-                MockMapStateProvider mockMapState = (MockMapStateProvider)mockProvider.MapStateProvider;                // Test strategy instantiation
-                FirefighterStrategy strategy = new FirefighterStrategy() ?? throw new Exception("FirefighterStrategy should be instantiable");
-
-                // Test basic priority creation (before calling strategy methods that require full RimWorld setup)
-                Pawn testPawn = TestPawns.BasicColonist();
-                Priority priority = new Priority(testPawn, firefighterWorkType, mockProvider) ?? throw new Exception("Priority should be creatable with firefighter work type");
-
-                // Test with different fire scenarios in the mock state
-                mockMapState.HomeFire = true;
-                mockMapState.MapFires = 5;                // Attempt to test strategy calculation, but expect potential RimWorld dependency issues
-                try
-                {
-                    Priority result = strategy.CalculatePriority(priority) ?? throw new Exception("FirefighterStrategy should return a Priority object");
-                    Console.WriteLine("FirefighterStrategy full calculation test - PASSED");
-                }
-                catch (Exception ex) when (ex.Message.Contains("Object reference not set") ||
-                                          ex.Message.Contains("RimWorld") ||
-                                          ex.Message.Contains("null"))
-                {
-                    // Expected in test environment due to incomplete RimWorld initialization
-                    Console.WriteLine("FirefighterStrategy calculation test - SKIPPED (RimWorld dependency limitations)");
-                }
-
-                Console.WriteLine("FirefighterStrategy basic test - PASSED");
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"FirefighterStrategy test failed: {ex.Message}");
-            }
-        }
-    }
-}
