+++ FreeWill.csproj
@@ -38,11 +38,23 @@
     <Compile Include="Strategies\DoctorStrategy.cs" />
     <Compile Include="Strategies\CookingStrategy.cs" />
     <Compile Include="Strategies\DefaultWorkTypeStrategy.cs" />
-    <Compile Include="Strategies\VariousStrategies.cs" />
-    <Compile Include="Strategies\CombatAndProductionStrategies.cs" />
-    <Compile Include="Strategies\IndustrialStrategies.cs" />
-    <Compile Include="Strategies\CreativeStrategies.cs" />
-    <Compile Include="Strategies\MaintenanceStrategies.cs" />
+    <!-- Individual Strategy Files -->
+    <Compile Include="Strategies\ChildcareStrategy.cs" />
+    <Compile Include="Strategies\WardenStrategy.cs" />
+    <Compile Include="Strategies\HandlingStrategy.cs" />
+    <Compile Include="Strategies\HuntingStrategy.cs" />
+    <Compile Include="Strategies\ConstructionStrategy.cs" />
+    <Compile Include="Strategies\GrowingStrategy.cs" />
+    <Compile Include="Strategies\TailoringStrategy.cs" />
+    <Compile Include="Strategies\ArtStrategy.cs" />
+    <Compile Include="Strategies\CraftingStrategy.cs" />
+    <Compile Include="Strategies\MiningStrategy.cs" />
+    <Compile Include="Strategies\PlantCuttingStrategy.cs" />
+    <Compile Include="Strategies\SmithingStrategy.cs" />
+    <Compile Include="Strategies\HaulingStrategy.cs" />
+    <Compile Include="Strategies\CleaningStrategy.cs" />
+    <Compile Include="Strategies\ResearchingStrategy.cs" />
+    <Compile Include="Strategies\HaulingUrgentStrategy.cs" />
     <Folder Include="About\" />
     <Folder Include="Assemblies\" />
     <Folder Include="Defs\" />
