--- Strategies/CombatAndProductionStrategies.cs
@@ -1,102 +1,0 @@
-using Verse;
-
-namespace FreeWill
-{
-    /// <summary>
-    /// Strategy for hunting work type priority calculation.
-    /// </summary>
-    public class HuntingStrategy : BaseWorkTypeStrategy
-    {
-        public override WorkTypeDef WorkType => GetWorkTypeDef("Hunting");
-
-        public override Priority CalculatePriority(Priority priority)
-        {
-            return priority
-                .ConsiderRelevantSkills()
-                .ConsiderCarryingCapacity()
-                .ConsiderIsAnyoneElseDoing()
-                .ConsiderBestAtDoing()
-                .ConsiderPassion()
-                .ConsiderThoughts()
-                .ConsiderInspiration()
-                .ConsiderLowFood(0.3f)
-                .ConsiderWeaponRange()
-                .ConsiderColonistLeftUnburied()
-                .ConsiderMovementSpeed()
-                .ConsiderHealth()
-                .ConsiderAteRawFood()
-                .ConsiderBored()
-                .ConsiderHasHuntingWeapon()
-                .ConsiderBrawlersNotHunting()
-                .ConsiderFire()
-                .ConsiderBuildingImmunity()
-                .ConsiderCompletingTask()
-                .ConsiderColonistsNeedingTreatment()
-                .ConsiderDownedColonists()
-                .ConsiderColonyPolicy();
-        }
-    }
-
-    /// <summary>
-    /// Strategy for construction work type priority calculation.
-    /// </summary>
-    public class ConstructionStrategy : BaseWorkTypeStrategy
-    {
-        public override WorkTypeDef WorkType => GetWorkTypeDef("Construction");
-
-        public override Priority CalculatePriority(Priority priority)
-        {
-            priority.ConsiderRelevantSkills()
-                .ConsiderCarryingCapacity()
-                .ConsiderIsAnyoneElseDoing()
-                .ConsiderBestAtDoing()
-                .ConsiderPassion()
-                .ConsiderThoughts()
-                .ConsiderInspiration()
-                .ConsiderAnimalPen();
-            _ = priority.ConsiderLowFood(-0.3f)
-                .ConsiderColonistLeftUnburied()
-                .ConsiderHealth()
-                .ConsiderAteRawFood()
-                .ConsiderBored()
-                .ConsiderFire()
-                .ConsiderBuildingImmunity()
-                .ConsiderCompletingTask()
-                .ConsiderColonistsNeedingTreatment()
-                .ConsiderDownedColonists()
-                .ConsiderColonyPolicy();
-            return priority;
-        }
-    }
-
-    /// <summary>
-    /// Strategy for growing work type priority calculation.
-    /// </summary>
-    public class GrowingStrategy : BaseWorkTypeStrategy
-    {
-        public override WorkTypeDef WorkType => GetWorkTypeDef("Growing");
-
-        public override Priority CalculatePriority(Priority priority)
-        {
-            return priority
-                .ConsiderRelevantSkills()
-                .ConsiderCarryingCapacity()
-                .ConsiderIsAnyoneElseDoing()
-                .ConsiderBestAtDoing()
-                .ConsiderPassion()
-                .ConsiderThoughts()
-                .ConsiderInspiration()
-                .ConsiderLowFood(0.3f)
-                .ConsiderColonistLeftUnburied()
-                .ConsiderHealth()
-                .ConsiderAteRawFood()
-                .ConsiderBored()
-                .ConsiderFire()
-                .ConsiderBuildingImmunity()
-                .ConsiderCompletingTask()
-                .ConsiderColonistsNeedingTreatment()
-                .ConsiderDownedColonists()
-                .ConsiderColonyPolicy();
-        }
-    }
-}
