--- Strategies/CreativeStrategies.cs
@@ -1,101 +1,0 @@
-using Verse;
-
-namespace FreeWill
-{
-    /// <summary>
-    /// Strategy for tailoring work type priority calculation.
-    /// </summary>
-    public class TailoringStrategy : BaseWorkTypeStrategy
-    {
-        public override WorkTypeDef WorkType => GetWorkTypeDef("Tailoring");
-
-        public override Priority CalculatePriority(Priority priority)
-        {
-            return priority
-                .ConsiderRelevantSkills()
-                .ConsiderCarryingCapacity()
-                .ConsiderIsAnyoneElseDoing()
-                .ConsiderBestAtDoing()
-                .ConsiderBeautyExpectations()
-                .ConsiderPassion()
-                .ConsiderThoughts()
-                .ConsiderInspiration()
-                .ConsiderLowFood(-0.3f)
-                .ConsiderNeedingWarmClothes()
-                .ConsiderColonistLeftUnburied()
-                .ConsiderHealth()
-                .ConsiderAteRawFood()
-                .ConsiderBored()
-                .ConsiderFire()
-                .ConsiderBuildingImmunity()
-                .ConsiderCompletingTask()
-                .ConsiderColonistsNeedingTreatment()
-                .ConsiderDownedColonists()
-                .ConsiderColonyPolicy();
-        }
-    }
-
-    /// <summary>
-    /// Strategy for art work type priority calculation.
-    /// </summary>
-    public class ArtStrategy : BaseWorkTypeStrategy
-    {
-        public override WorkTypeDef WorkType => GetWorkTypeDef("Art");
-
-        public override Priority CalculatePriority(Priority priority)
-        {
-            return priority
-                .ConsiderRelevantSkills()
-                .ConsiderCarryingCapacity()
-                .ConsiderIsAnyoneElseDoing()
-                .ConsiderBestAtDoing()
-                .ConsiderBeautyExpectations()
-                .ConsiderPassion()
-                .ConsiderThoughts()
-                .ConsiderInspiration()
-                .ConsiderLowFood(-0.3f)
-                .ConsiderColonistLeftUnburied()
-                .ConsiderHealth()
-                .ConsiderAteRawFood()
-                .ConsiderBored()
-                .ConsiderFire()
-                .ConsiderBuildingImmunity()
-                .ConsiderCompletingTask()
-                .ConsiderColonistsNeedingTreatment()
-                .ConsiderDownedColonists()
-                .ConsiderColonyPolicy();
-        }
-    }
-
-    /// <summary>
-    /// Strategy for crafting work type priority calculation.
-    /// </summary>
-    public class CraftingStrategy : BaseWorkTypeStrategy
-    {
-        public override WorkTypeDef WorkType => GetWorkTypeDef("Crafting");
-
-        public override Priority CalculatePriority(Priority priority)
-        {
-            return priority
-                .ConsiderRelevantSkills()
-                .ConsiderCarryingCapacity()
-                .ConsiderIsAnyoneElseDoing()
-                .ConsiderBestAtDoing()
-                .ConsiderBeautyExpectations()
-                .ConsiderPassion()
-                .ConsiderThoughts()
-                .ConsiderInspiration()
-                .ConsiderLowFood(-0.3f)
-                .ConsiderColonistLeftUnburied()
-                .ConsiderHealth()
-                .ConsiderAteRawFood()
-                .ConsiderBored()
-                .ConsiderFire()
-                .ConsiderBuildingImmunity()
-                .ConsiderCompletingTask()
-                .ConsiderColonistsNeedingTreatment()
-                .ConsiderDownedColonists()
-                .ConsiderColonyPolicy();
-        }
-    }
-}
