+++ Strategies/GrowingStrategy.cs
@@ -1,0 +1,38 @@
+using Verse;
+
+namespace FreeWill
+{
+    /// <summary>
+    /// Strategy for growing work type priority calculation.
+    /// Growing involves planting, tending, and harvesting crops for food production.
+    /// Priority significantly increases during food shortages as crops are essential for survival.
+    /// Considers agricultural skills and long-term colony food security.
+    /// </summary>
+    public class GrowingStrategy : BaseWorkTypeStrategy
+    {
+        public override WorkTypeDef WorkType => GetWorkTypeDef("Growing");
+
+        public override Priority CalculatePriority(Priority priority)
+        {
+            return priority
+                .ConsiderRelevantSkills()
+                .ConsiderCarryingCapacity()
+                .ConsiderIsAnyoneElseDoing()
+                .ConsiderBestAtDoing()
+                .ConsiderPassion()
+                .ConsiderThoughts()
+                .ConsiderInspiration()
+                .ConsiderLowFood(0.3f) // Growing becomes critical when food is scarce
+                .ConsiderColonistLeftUnburied()
+                .ConsiderHealth()
+                .ConsiderAteRawFood()
+                .ConsiderBored()
+                .ConsiderFire()
+                .ConsiderBuildingImmunity()
+                .ConsiderCompletingTask()
+                .ConsiderColonistsNeedingTreatment()
+                .ConsiderDownedColonists()
+                .ConsiderColonyPolicy();
+        }
+    }
+}
