--- Strategies/IndustrialStrategies.cs
@@ -1,99 +1,0 @@
-using Verse;
-
-namespace FreeWill
-{
-    /// <summary>
-    /// Strategy for mining work type priority calculation.
-    /// </summary>
-    public class MiningStrategy : BaseWorkTypeStrategy
-    {
-        public override WorkTypeDef WorkType => GetWorkTypeDef("Mining");
-
-        public override Priority CalculatePriority(Priority priority)
-        {
-            return priority
-                .ConsiderRelevantSkills()
-                .ConsiderCarryingCapacity()
-                .ConsiderIsAnyoneElseDoing()
-                .ConsiderBestAtDoing()
-                .ConsiderPassion()
-                .ConsiderThoughts()
-                .ConsiderInspiration()
-                .ConsiderLowFood(-0.3f)
-                .ConsiderColonistLeftUnburied()
-                .ConsiderHealth()
-                .ConsiderAteRawFood()
-                .ConsiderBored()
-                .ConsiderFire()
-                .ConsiderBuildingImmunity()
-                .ConsiderCompletingTask()
-                .ConsiderColonistsNeedingTreatment()
-                .ConsiderDownedColonists()
-                .ConsiderColonyPolicy();
-        }
-    }
-
-    /// <summary>
-    /// Strategy for plant cutting work type priority calculation.
-    /// </summary>
-    public class PlantCuttingStrategy : BaseWorkTypeStrategy
-    {
-        public override WorkTypeDef WorkType => GetWorkTypeDef("PlantCutting");
-
-        public override Priority CalculatePriority(Priority priority)
-        {
-            return priority
-                .ConsiderRelevantSkills()
-                .ConsiderIsAnyoneElseDoing()
-                .ConsiderBestAtDoing()
-                .ConsiderPassion()
-                .ConsiderThoughts()
-                .ConsiderInspiration()
-                .ConsiderGauranlenPruning()
-                .ConsiderLowFood(0.3f)
-                .ConsiderHealth()
-                .ConsiderPlantsBlighted()
-                .ConsiderBored()
-                .ConsiderFire()
-                .ConsiderBuildingImmunity()
-                .ConsiderCompletingTask()
-                .ConsiderColonistsNeedingTreatment()
-                .ConsiderDownedColonists()
-                .ConsiderColonyPolicy();
-        }
-    }
-
-    /// <summary>
-    /// Strategy for smithing work type priority calculation.
-    /// </summary>
-    public class SmithingStrategy : BaseWorkTypeStrategy
-    {
-        public override WorkTypeDef WorkType => GetWorkTypeDef("Smithing");
-
-        public override Priority CalculatePriority(Priority priority)
-        {
-            return priority
-                .ConsiderRelevantSkills()
-                .ConsiderCarryingCapacity()
-                .ConsiderFinishedMechGestators()
-                .ConsiderRepairingMech()
-                .ConsiderIsAnyoneElseDoing()
-                .ConsiderBestAtDoing()
-                .ConsiderBeautyExpectations()
-                .ConsiderPassion()
-                .ConsiderThoughts()
-                .ConsiderInspiration()
-                .ConsiderLowFood(-0.3f)
-                .ConsiderColonistLeftUnburied()
-                .ConsiderHealth()
-                .ConsiderAteRawFood()
-                .ConsiderBored()
-                .ConsiderFire()
-                .ConsiderBuildingImmunity()
-                .ConsiderCompletingTask()
-                .ConsiderColonistsNeedingTreatment()
-                .ConsiderDownedColonists()
-                .ConsiderColonyPolicy();
-        }
-    }
-}
