+++ Strategies/MiningStrategy.cs
@@ -1,0 +1,38 @@
+using Verse;
+
+namespace FreeWill
+{
+    /// <summary>
+    /// Strategy for mining work type priority calculation.
+    /// Mining involves extracting valuable resources from stone chunks and underground deposits.
+    /// Priority decreases during food shortages as mining is less critical than immediate survival needs.
+    /// Considers mining skills and the strategic value of different mineral resources.
+    /// </summary>
+    public class MiningStrategy : BaseWorkTypeStrategy
+    {
+        public override WorkTypeDef WorkType => GetWorkTypeDef("Mining");
+
+        public override Priority CalculatePriority(Priority priority)
+        {
+            return priority
+                .ConsiderRelevantSkills()
+                .ConsiderCarryingCapacity()
+                .ConsiderIsAnyoneElseDoing()
+                .ConsiderBestAtDoing()
+                .ConsiderPassion()
+                .ConsiderThoughts()
+                .ConsiderInspiration()
+                .ConsiderLowFood(-0.3f) // Mining becomes less important when food is scarce
+                .ConsiderColonistLeftUnburied()
+                .ConsiderHealth()
+                .ConsiderAteRawFood()
+                .ConsiderBored()
+                .ConsiderFire()
+                .ConsiderBuildingImmunity()
+                .ConsiderCompletingTask()
+                .ConsiderColonistsNeedingTreatment()
+                .ConsiderDownedColonists()
+                .ConsiderColonyPolicy();
+        }
+    }
+}
