--- Strategies/VariousStrategies.cs
@@ -1,95 +1,0 @@
-using Verse;
-
-namespace FreeWill
-{
-    /// <summary>
-    /// Strategy for childcare work type priority calculation.
-    /// </summary>
-    public class ChildcareStrategy : BaseWorkTypeStrategy
-    {
-        public override WorkTypeDef WorkType => GetWorkTypeDef("Childcare");
-
-        public override Priority CalculatePriority(Priority priority)
-        {
-            priority.Set(0.5f, "FreeWillPriorityChildcareDefault".TranslateSimple);
-            return priority
-                .ConsiderRelevantSkills(shouldAdd: true)
-                .ConsiderPassion()
-                .ConsiderThoughts()
-                .ConsiderInspiration()
-                .ConsiderHealth()
-                .ConsiderAteRawFood()
-                .ConsiderBored()
-                .ConsiderFire()
-                .ConsiderBuildingImmunity()
-                .ConsiderCompletingTask()
-                .ConsiderColonistsNeedingTreatment()
-                .ConsiderDownedColonists()
-                .ConsiderColonyPolicy();
-        }
-    }
-
-    /// <summary>
-    /// Strategy for warden work type priority calculation.
-    /// </summary>
-    public class WardenStrategy : BaseWorkTypeStrategy
-    {
-        public override WorkTypeDef WorkType => GetWorkTypeDef("Warden");
-
-        public override Priority CalculatePriority(Priority priority)
-        {
-            return priority
-                .ConsiderRelevantSkills()
-                .ConsiderCarryingCapacity()
-                .ConsiderIsAnyoneElseDoing()
-                .ConsiderBestAtDoing()
-                .ConsiderPassion()
-                .ConsiderThoughts()
-                .ConsiderInspiration()
-                .ConsiderLowFood(-0.3f)
-                .ConsiderSuppressionNeed()
-                .ConsiderColonistLeftUnburied()
-                .ConsiderHealth()
-                .ConsiderAteRawFood()
-                .ConsiderBored()
-                .ConsiderFire()
-                .ConsiderBuildingImmunity()
-                .ConsiderCompletingTask()
-                .ConsiderColonistsNeedingTreatment()
-                .ConsiderDownedColonists()
-                .ConsiderColonyPolicy();
-        }
-    }
-
-    /// <summary>
-    /// Strategy for animal handling work type priority calculation.
-    /// </summary>
-    public class HandlingStrategy : BaseWorkTypeStrategy
-    {
-        public override WorkTypeDef WorkType => GetWorkTypeDef("Handling");
-
-        public override Priority CalculatePriority(Priority priority)
-        {
-            return priority
-                .ConsiderRelevantSkills()
-                .ConsiderCarryingCapacity()
-                .ConsiderIsAnyoneElseDoing()
-                .ConsiderBestAtDoing()
-                .ConsiderPassion()
-                .ConsiderThoughts()
-                .ConsiderInspiration()
-                .ConsiderAnimalsRoaming()
-                .ConsiderColonistLeftUnburied()
-                .ConsiderMovementSpeed()
-                .ConsiderHealth()
-                .ConsiderAteRawFood()
-                .ConsiderBored()
-                .ConsiderFire()
-                .ConsiderBuildingImmunity()
-                .ConsiderCompletingTask()
-                .ConsiderColonistsNeedingTreatment()
-                .ConsiderDownedColonists()
-                .ConsiderColonyPolicy();
-        }
-    }
-}
